requirements.txt
openai>=1.50.0
PyGithub>=2.3.0
